scylla-artifacts: Scan cassandra-stress output for java exceptions
cassandra-stress will happily ignore exceptions that happened
inside threads and return exit code !=0. That is of course
bad for an automated test perspective, given that this condition
will yield a false negative. Let's be safe and scan the test
output for such conditions, and fail the test in case they are
found.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>