scylla-artifacts.py: Refactor package install code
The current version of the test is cluttered and could
use some logical separation of code responsible for
installing Scylla and actual test of the installed artifacts.

So let's do that by splitting the code into installers,
service managers and the test itself. This also updates
the release mode testing to use the new repository files
for Fedora, CentOS and Ubuntu.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>